pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "google-genai>=1.49.0",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
]
